local_shaots_backend/requirements.txt
bcrypt==4.1.1
bleak==0.21.1
blinker==1.7.0
boto3==1.34.0
botocore==1.34.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dbus-fast==2.22.1
docopt==0.6.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
num2words==0.5.13
packaging==23.2
pdfkit==1.0.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.4.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
smart-open==7.0.4
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.16.0
yarg==0.1.9
qrcode==7.4.2